refactor: добавил возможность реэкспорта компонент
