+++ src/components/LangSelect/index.js
@@ -1,0 +1,1 @@
+export * from './LangSelect';
